// backend/src/modules/ai/types/action.types.ts
export enum AIAction {
  NONE = 'NONE',
  SEARCH_CUSTOMER = 'SEARCH_CUSTOMER',
  GET_CUSTOMER = 'GET_CUSTOMER',
  GET_SERVICES = 'GET_SERVICES',
  GET_SERVICE_DETAILS = 'GET_SERVICE_DETAILS',
  GET_STAFF = 'GET_STAFF',
  GET_STAFF_DETAILS = 'GET_STAFF_DETAILS',
  CHECK_AVAILABILITY = 'CHECK_AVAILABILITY',
  CREATE_APPOINTMENT = 'CREATE_APPOINTMENT',
  GET_APPOINTMENTS = 'GET_APPOINTMENTS',
  CANCEL_APPOINTMENT = 'CANCEL_APPOINTMENT',
  RESCHEDULE_APPOINTMENT = 'RESCHEDULE_APPOINTMENT',
  GET_BUSINESS_INFO = 'GET_BUSINESS_INFO',
  GET_CLINIC_DIRECTIONS = 'GET_CLINIC_DIRECTIONS',
  GET_CABIN_LOCATION = 'GET_CABIN_LOCATION',
  GET_WAITING_POLICY = 'GET_WAITING_POLICY',
  TRIAGE_SYMPTOM = 'TRIAGE_SYMPTOM',
  EMERGENCY_ESCALATION = 'EMERGENCY_ESCALATION',
  ESCALATE_EMERGENCY = 'EMERGENCY_ESCALATION',
}